+++ src/scg_optimal_method.c
@@ -67,7 +67,8 @@ int igraph_i_optimal_partition(const igraph_real_t *v, int *gr, int n,
     vs[i].ind = i;
   }
   
-  qsort(vs, n, sizeof(igraph_i_scg_indval_t), igraph_i_compare_ind_val);
+  qsort(vs, (size_t) n, sizeof(igraph_i_scg_indval_t), 
+	igraph_i_compare_ind_val);
   
   non_ties = 1;
   for (i=1; i<n; i++) {
